requirements.txt
dspy-ai
openai
roonapi
pydantic
python-dotenv
rich
# Optional if you want to try other LMs
# ollama
# litellm